+++ csql/__init__.py
@@ -1,9 +1,12 @@
 # anything exported through here is considered public api.
 
-from .api import Q as Q
-from .models.query import (
-	Query as Query,
-	Parameters as Parameters,
-	ParameterPlaceholder as ParameterPlaceholder,
-	RenderedQuery as RenderedQuery
-)
+from ._.api import *
+from ._.models.query import *
+
+__all__ = [
+	'Q',
+	'Query',
+	'Parameters',
+	'ParameterPlaceholder',
+	'RenderedQuery'
+]
