+++ csql/dialect.py
@@ -1,4 +1,4 @@
-from .models.dialect import (
+from ._.models.dialect import (
 	SQLDialect as SQLDialect,
 	ParamStyle as ParamStyle,
 	DefaultDialect as DefaultDialect,
